+++ jabgui/src/main/java/org/jabref/gui/newentry/NewEntryViewModel.java
@@ -32,13 +32,14 @@ import org.jabref.logic.importer.FetcherServerException;
 import org.jabref.logic.importer.IdBasedFetcher;
 import org.jabref.logic.importer.ParseException;
 import org.jabref.logic.importer.WebFetchers;
+import org.jabref.logic.importer.fileformat.BibliographyFromPdfImporter;
 import org.jabref.logic.importer.fileformat.BibtexParser;
 import org.jabref.logic.importer.plaincitation.GrobidPlainCitationParser;
 import org.jabref.logic.importer.plaincitation.LlmPlainCitationParser;
+import org.jabref.logic.importer.plaincitation.MultiplePlainCitationsParser;
 import org.jabref.logic.importer.plaincitation.PlainCitationParser;
 import org.jabref.logic.importer.plaincitation.PlainCitationParserChoice;
 import org.jabref.logic.importer.plaincitation.RuleBasedPlainCitationParser;
-import org.jabref.logic.importer.plaincitation.MultiplePlainCitationsParser;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.logic.layout.LayoutFormatter;
 import org.jabref.logic.layout.format.DOIStrip;
@@ -357,8 +358,10 @@ public class NewEntryViewModel {
             }
 
             final PlainCitationParser parser = switch (parserChoice) {
-                case PlainCitationParserChoice.RULE_BASED ->
+                case PlainCitationParserChoice.RULE_BASED_SIMPLE ->
                         new RuleBasedPlainCitationParser();
+                case PlainCitationParserChoice.RULE_BASED_IEEE ->
+                        new BibliographyFromPdfImporter(preferences.getCitationKeyPatternPreferences());
                 case PlainCitationParserChoice.GROBID ->
                         new GrobidPlainCitationParser(preferences.getGrobidPreferences(), preferences.getImportFormatPreferences());
                 case PlainCitationParserChoice.LLM ->